Remove cli_common: merge remaining code directly into plot_heatmap
