+++ src/evaluating_rewards/analysis/dissimilarity_heatmaps/plot_heatmap.py
@@ -27,7 +27,7 @@ import sacred
 
 from evaluating_rewards import serialize
 from evaluating_rewards.analysis import stylesheets, visualize
-from evaluating_rewards.analysis.dissimilarity_heatmaps import cli_common, heatmaps
+from evaluating_rewards.analysis.dissimilarity_heatmaps import heatmaps, reward_masks
 from evaluating_rewards.distances import transitions_datasets
 from evaluating_rewards.scripts import script_utils
 from evaluating_rewards.scripts.distance import common
@@ -36,8 +36,9 @@ plot_heatmap_ex = sacred.Experiment("plot_heatmap")
 logger = logging.getLogger("evaluating_rewards.analysis.plot_heatmap")
 
 
-cli_common.make_config(plot_heatmap_ex)
 transitions_datasets.make_config(plot_heatmap_ex)
+
+# Default configs (always applied.)
 
 
 @plot_heatmap_ex.config
@@ -70,6 +71,38 @@ def logging_config(env_name, log_root):
     del _
 
 
+@plot_heatmap_ex.config
+def figure_config():
+    """Defaults for figure parameters."""
+    heatmap_kwargs = {
+        "masks": {"all": [reward_masks.always_true]},
+        "after_plot": heatmaps.horizontal_ticks,
+    }
+    styles = ["paper", "heatmap", "heatmap-3col", "tex"]
+    styles_for_env = []
+    save_kwargs = {
+        "fmt": "pdf",
+    }
+    _ = locals()
+    del _
+
+
+# General formatting instructions.
+# For example, how large a figure? What precision?
+
+
+@plot_heatmap_ex.named_config
+def large():
+    """Large output size, high precision."""
+    styles = ["paper", "heatmap", "heatmap-1col", "tex"]
+    heatmap_kwargs = {
+        "fmt": heatmaps.short_e,
+        "cbar_kws": dict(fraction=0.05),
+    }
+    _ = locals()
+    del _
+
+
 @plot_heatmap_ex.named_config
 def paper():
     """Figures suitable for inclusion in paper."""
@@ -86,6 +119,90 @@ def test():
     """Intended for debugging/unit test."""
     # Do not include "tex" in styles here: this will break on CI.
     styles = ["paper", "heatmap-1col"]
+    _ = locals()
+    del _
+
+
+# Environment specific configs.
+# (Mostly special-case formatting to make presentation clearer.)
+
+
+def _norm(args: Iterable[str]) -> bool:
+    return any(reward_masks.match("evaluating_rewards/PointMassGroundTruth-v0")(args))
+
+
+def _hopper_activity(args: Iterable[str]) -> bool:
+    pattern = r"evaluating_rewards/(.*)(GroundTruth|Backflip)(.*)"
+    repl = reward_masks.replace(pattern, r"\1\2")(args)
+    return len(set(repl)) > 1 and reward_masks.no_ctrl(args)
+
+
+@plot_heatmap_ex.named_config
+def point_mass():
+    """Heatmaps for evaluating_rewards/PointMass* environments."""
+    locals().update(**common.COMMON_CONFIGS["point_mass"])
+    heatmap_kwargs = {}
+    heatmap_kwargs["masks"] = {
+        "diagonal": [reward_masks.zero, reward_masks.same],
+        "control": [reward_masks.zero, reward_masks.control],
+        "dense_vs_sparse": [reward_masks.zero, reward_masks.sparse_or_dense],
+        "norm": [reward_masks.zero, reward_masks.same, _norm],
+        "all": [reward_masks.always_true],
+    }
+    _ = locals()
+    del _
+
+
+@plot_heatmap_ex.named_config
+def point_maze():
+    """Heatmaps for imitation/PointMaze{Left,Right}-v0 environments."""
+    locals().update(**common.COMMON_CONFIGS["point_maze"])
+    heatmap_kwargs = {
+        "masks": {"all": [reward_masks.always_true]},  # "all" is still only 2x2
+    }
+    _ = locals()
+    del _
+
+
+@plot_heatmap_ex.named_config
+def point_maze_learned():
+    """Compare rewards learned in PointMaze to the ground-truth reward."""
+    # Analyzes models generated by `runners/transfer_point_maze.sh`.
+    locals().update(**common.COMMON_CONFIGS["point_maze_learned"])
+
+
+@plot_heatmap_ex.named_config
+def half_cheetah():
+    """Heatmaps for seals/HalfCheetah-v0."""
+    locals().update(**common.COMMON_CONFIGS["half_cheetah"])
+    heatmap_kwargs = {
+        "masks": {
+            "diagonal": [reward_masks.zero, reward_masks.same],
+            "control": [reward_masks.zero, reward_masks.control],
+            "direction": [reward_masks.zero, reward_masks.direction],
+            "no_ctrl": [reward_masks.zero, reward_masks.no_ctrl],
+            "all": [reward_masks.always_true],
+        },
+    }
+    styles_for_env = ["small-labels"]  # downscale emoji labels slightly
+    _ = locals()
+    del _
+
+
+@plot_heatmap_ex.named_config
+def hopper():
+    """Heatmaps for seals/Hopper-v0."""
+    locals().update(**common.COMMON_CONFIGS["hopper"])
+    heatmap_kwargs = {}
+    heatmap_kwargs["masks"] = {
+        "diagonal": [reward_masks.zero, reward_masks.same],
+        "control": [reward_masks.zero, reward_masks.control],
+        "direction": [reward_masks.zero, reward_masks.direction],
+        "no_ctrl": [reward_masks.zero, reward_masks.no_ctrl],
+        "different_activity": [reward_masks.zero, _hopper_activity],
+        "all": [reward_masks.always_true],
+    }
+    styles_for_env = ["tiny-font"]
     _ = locals()
     del _
 
@@ -156,21 +273,6 @@ def oned_mapping_to_series(
     return vals
 
 
-def twod_mapping_to_multi_series(
-    aggregated: Mapping[str, Mapping[Any, float]]
-) -> Mapping[str, pd.Series]:
-    """Converts a nested mapping to a mapping of dissimilarity series.
-
-    Args:
-        aggregated: A mapping over a mapping from strings to sequences of floats.
-
-    Returns:
-        A mapping from strings to MultiIndex series returned by `oned_mapping_to_series`,
-        after transposing the inner and outer keys of the mapping.
-    """
-    return {k: oned_mapping_to_series(v) for k, v in aggregated.items()}
-
-
 def select_subset(
     vals: common.AggregatedDistanceReturn,
     x_reward_cfgs: Iterable[common.RewardCfg],
@@ -195,16 +297,30 @@ def plot_heatmap(
     heatmap_kwargs: Mapping[str, Any],
     save_kwargs: Mapping[str, Any],
 ) -> None:
-    """Plots a figure for each entry loaded from `vals_path`."""
-    # TODO(adam): document args in docstring
+    """Plots a figure for each entry loaded from `vals_path`.
+
+    Args:
+        vals_path: path to pickle file containing aggregated values.
+            Produced by `evaluating_rewards.scripts.distance.*`.
+        data_root: the root with respect to canonicalize reward configurations.
+        x_reward_cfgs: tuples of reward_type and reward_path for x-axis.
+        y_reward_cfgs: tuples of reward_type and reward_path for y-axis.
+        styles: styles to apply from `evaluating_rewards.analysis.stylesheets`.
+        styles_for_env: extra styles to apply, concatenated with above.
+        log_dir: directory to save data to.
+        timestamp: timestamp + unique identifier, usually a component of `log_dir`.
+        heatmap_kwargs: passed through to `heatmaps.compact_heatmaps`.
+        save_kwargs: passed through to `analysis.save_figs`.
+    """
     # Sacred turns our tuples into lists :(, undo
     x_reward_cfgs = [common.canonicalize_reward_cfg(cfg, data_root) for cfg in x_reward_cfgs]
     y_reward_cfgs = [common.canonicalize_reward_cfg(cfg, data_root) for cfg in y_reward_cfgs]
+
     # TODO(adam): how to specify vals_path?
     with open(vals_path, "rb") as f:
         aggregated_raw = pickle.load(f)
     aggregated_raw = select_subset(aggregated_raw, x_reward_cfgs, y_reward_cfgs)
-    aggregated = twod_mapping_to_multi_series(aggregated_raw)
+    aggregated = {k: oned_mapping_to_series(v) for k, v in aggregated_raw.items()}
 
     logging.info("Plotting figures")
     vals_dir = os.path.dirname(vals_path)
